refactor and add new endpoints

+++ api/src/courses_api.py
@@ -43,8 +43,10 @@ def update_course(db: Session, id: int, course: schema.UpdateCourseModel):
                 db_models.Course.hashtags: course.hashtags,
                 db_models.Course.type: course.type,
                 db_models.Course.exams: course.exams,
-                db_models.Course.suscription: course.suscription,
-                db_models.Course.location: course.location
+                db_models.Course.suscription: course.subscription,
+                db_models.Course.location: course.location,
+                db_models.Course.conditions: course.enrollment_conditions,
+                db_models.Course.unenrollment_conditions: course.unenrollment_conditions
             }
         )
         db.commit()
@@ -55,53 +57,46 @@ def update_course(db: Session, id: int, course: schema.UpdateCourseModel):
     return response
 
 
-def get_courses(db: Session):
-    courses_data = db.query(db_models.Course).all()
-    if not courses_data:
+def get_courses(db: Session, id, category, subscription, creator):
+    query = db.query(db_models.Course)
+    if id:
+        query = query.filter(db_models.Course.id == id)
+    if creator:
+        query = query.filter(db_models.Course.creator == creator)
+    if category:
+        query = query.filter(db_models.Course.category == category)
+    if subscription:
+        query = query.filter(db_models.Course.subscription == subscription)
+    courses = query.all()
+    if not courses:
         return create_message_response("No existen cursos")
-    return get_courses_dict(courses_data)
+    return courses
 
 
-def get_courses_by_suscription(db: Session, suscription):
-    courses_data = db.query(db_models.Course).filter(db_models.Course.suscription == suscription).all()
-    if not courses_data:
-        return create_message_response("No existen cursos para el tipo de suscripción " + suscription)
-    return get_courses_dict(courses_data)
-
-
-def get_courses_by_category(db: Session, category):
-    courses_data = db.query(db_models.Course).filter(db_models.Course.category == category).all()
-    if not courses_data:
-        return create_message_response("No existen cursos para la cateogría " + category)
-    return get_courses_dict(courses_data)
-
-
-def get_courses_by_creator(db: Session, creator_id):
-    courses_data = db.query(db_models.Course).filter(db_models.Course.creator == creator_id).all()
-    if not courses_data:
-        return create_message_response("Aún no ha creado ningún curso")
-    return get_courses_dict(courses_data)
-
-
-def add_collaborator(db: Session, course_id, collaborator_id):
-    if course_id=="" or collaborator_id=="":
-        raise InvalidOperationException("Uno o mas campos se encuentran vacíos")
-    collaborator_model = db_models.Collaborators(course_id = course_id, collaborator_id=collaborator_id)
-    db.add(collaborator_model)
+def add_collaborator(db: Session, course_id, collaborator_email):
+    creator = db.query(db_models.User).join(db_models.Course).filter(db_models.Course.id == course_id, db_models.User.id == db_models.Course.creator).first()
+    if not creator:
+        raise InvalidOperationException("El curso no existe")
+    if creator.email == collaborator_email:
+        raise InvalidOperationException("No se puede añadir como colaborador al usuario creador del curso")
+    else:
+       user= db.query(db_models.User).filter(db_models.User.email == collaborator_email).first()
+       if not user:
+           raise InvalidOperationException("El usuario no existe")
+       collaborator_model = db_models.Collaborators(course_id=course_id, collaborator_id=user.id)
+       db.add(collaborator_model)
     try:
         db.commit()
     except sqlalchemy.exc.IntegrityError as e:
-        db.rollback()
         if e.orig.pgcode == '23505':
             raise InvalidOperationException("El colaborador ya se encuentra asociado al curso.")
-        else:
-            raise InvalidOperationException("Uno o mas campos son incorrectos")
     response = create_message_response("El colaborador se agregó correctamente")
     return response
 
 
 def get_collaborators_by_course(db: Session, course_id):
-    data = db.query(db_models.Collaborators.collaborator_id, db_models.Users.name + " " + db_models.Users.lastname).join(db_models.Users).filter\
+    data = db.query(db_models.Collaborators.collaborator_id, db_models.User.name + " " + db_models.User.lastname).join(
+        db_models.User).filter \
         (db_models.Collaborators.course_id == course_id).all()
     if not data:
         return create_message_response("No Hay colaboradores asociados al curso")
@@ -113,9 +108,9 @@ def get_collaborators_by_course(db: Session, course_id):
     return collaborators
 
 
-def get_courses_by_collaborator(db: Session, collaborator_id):
-    data = db.query(db_models.Collaborators.course_id, db_models.Course.title).join(db_models.Course).filter\
-        (db_models.Collaborators.collaborator_id == collaborator_id).all()
+def get_courses_by_collaborator(db: Session, collaborator_email):
+    data = db.query(db_models.Course.id,db_models.Course.title).join(db_models.Collaborators).join(db_models.User).filter \
+        (db_models.User.email == collaborator_email).all()
     if not data:
         return create_message_response("El colaborador no se encuentra asociado a ningun curso")
     courses = {}
@@ -126,10 +121,68 @@ def get_courses_by_collaborator(db: Session, collaborator_id):
     return courses
 
 
-def remove_collaborator(db:Session, course_id, collaborator_id):
-    result = db.query(db_models.Collaborators).filter(db_models.Collaborators.collaborator_id == collaborator_id).filter(
+def remove_collaborator(db: Session, course_id, collaborator_email):
+    collaborator = db.query(db_models.User).filter(db_models.User.email==collaborator_email).first()
+    if not collaborator:
+        raise InvalidOperationException("El colaborador no existe")
+    result = db.query(db_models.Collaborators).filter(
+        db_models.Collaborators.collaborator_id == collaborator.id).filter(
         db_models.Collaborators.course_id == course_id).delete()
+    if not result:
+        raise InvalidOperationException("El colaborador no se encuentra asociado al curso")
+    db.commit()
+    return create_message_response("El colaborador se ha removido con éxito")
+
+
+def add_to_favorites(db: Session, favorite: schema.FavoriteModel):
+    favorite_model = db_models.Favorite(student_id=favorite.student_id, course_id=favorite.course_id)
+    db.add(favorite_model)
+    try:
+        db.commit()
+    except sqlalchemy.exc.IntegrityError as e:
+        db.rollback()
+        if e.orig.pgcode == '23505':
+            raise InvalidOperationException("El alumno ya ha marcado el curso como favorito")
+        else:
+            raise InvalidOperationException("Uno o mas campos son incorrectos")
+    response = create_message_response("Se agregó el curso a favoritos")
+    return response
+
+
+def delete_favorite(db: Session, favorite: schema.FavoriteModel):
+    result = db.query(db_models.Favorite).filter(
+        db_models.Favorite.student_id == favorite.student_id).filter(
+        db_models.Favorite.course_id == favorite.course_id).delete()
     if not result:
         raise InvalidOperationException("La operación no es válida")
     db.commit()
-    return create_message_response("El colaborador se ha removido con éxito")
+    return create_message_response("El curso ha sido removido de favoritos")
+
+
+def get_favorites(db, student_id, category, subscription):
+    query= db.query(db_models.Course).join(db_models.Favorite).filter(
+    db_models.Favorite.student_id == student_id).filter(db_models.Favorite.course_id == db_models.Course.id)
+    if category:
+        query = query.filter(db_models.Course.category == category)
+    if subscription:
+        query = query.filter(db_models.Course.subscription == subscription)
+    favorites = query.all()
+    if not favorites:
+        return create_message_response('El alumno no tiene cursos favoritos')
+    return favorites
+
+
+def get_subscriptions(db):
+    subscritpions = db.query(db_models.Subscription).all()
+    response = []
+    for subscription in subscritpions:
+        response.append(subscription.id)
+    return response
+
+
+def get_categories(db):
+    categories = db.query(db_models.Category).all()
+    response = []
+    for category in categories:
+        response.append(category.id)
+    return response
